+++ src/main/app/IO.java
@@ -1,3 +1,5 @@
+package app;
+
 // Bibliotecas
 import java.io.BufferedReader;
 import java.io.IOException;
